requirements.txt
streamlit==1.11.1
mysql-connector-python